requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
pymongo
python-multipart